[#11] Extract home ui into smaller widgets: SurveyPageViewer, HomeHeader and HomeFooter
